git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@5533 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ iomapserialize.cpp
@@ -273,8 +273,8 @@ bool IOMapSerialize::saveHouse(Database* db, House* house)
 
 		query.str("");
 		doorIds.insert(door->getDoorId());
-		query << house->getId() << ", " << g_config.getNumber(ConfigManager::WORLD_ID) << ", "
-			<< door->getDoorId() << ", " << db->escapeString(listText);
+		query << house->getId() << ", " << g_config.getNumber(ConfigManager::WORLD_ID)
+			<< ", " << (int32_t)door->getDoorId() << ", " << db->escapeString(listText);
 		if(!queryInsert.addRow(query.str()))
 			return false;
 	}
